refactor lift-job for easier understanding
